removed complete.js
