+++ webpack.config.babel.js
@@ -170,7 +170,6 @@ module.exports = [
 	{
 		entry: {
 			cmp: './index.js',
-			'cmp.complete': './complete.js',
 			portal: './docs/assets/portal.js'
 		},
